Fix hidden compiler warnings
... which were hidden due to suppressions from public cflags of the
protobuf library, which actually should have been private-cflags.

+++ src/buildtool/execution_api/execution_service/bytestream_server.cpp
@@ -91,7 +91,8 @@ auto BytestreamServiceImpl::Read(
     }
 
     ::google::bytestream::ReadResponse response;
-    for (auto it = to_read->make_iterator(request->read_offset());
+    for (auto it = to_read->make_iterator(
+             gsl::narrow<std::size_t>(request->read_offset()));
          it != to_read->end();
          ++it) {
         auto const chunk = *it;
